Make KV revisions opaque strings

+++ driver/sql/postgres/pgkv/keyspace.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 
 	"github.com/dogmatiq/persistencekit/driver/sql/postgres/internal/bigint"
+	"github.com/dogmatiq/persistencekit/internal/kvrevision"
 	"github.com/dogmatiq/persistencekit/kv"
 )
 
@@ -22,7 +23,7 @@ func (ks *keyspace) Name() string {
 func (ks *keyspace) Get(ctx context.Context, k []byte) (v []byte, r kv.Revision, err error) {
 	row := ks.db.QueryRowContext(
 		ctx,
-		`SELECT value, encoded_revision
+		`SELECT value, encoded_generation
 		FROM persistencekit.keyspace_pair
 		WHERE keyspace_id = $1
 		AND key = $2`,
@@ -30,17 +31,18 @@ func (ks *keyspace) Get(ctx context.Context, k []byte) (v []byte, r kv.Revision,
 		k,
 	)
 
+	var gen uint64
 	if err := row.Scan(
 		&v,
-		bigint.ConvertUnsigned(&r),
+		bigint.ConvertUnsigned(&gen),
 	); err != nil {
 		if err == sql.ErrNoRows {
-			return nil, 0, nil
+			return nil, "", nil
 		}
-		return nil, 0, fmt.Errorf("cannot scan keyspace pair: %w", err)
+		return nil, "", fmt.Errorf("cannot scan keyspace pair: %w", err)
 	}
 
-	return v, r, nil
+	return v, kvrevision.MarshalGeneration(gen), nil
 }
 
 func (ks *keyspace) Has(ctx context.Context, k []byte) (bool, error) {
@@ -62,47 +64,54 @@ func (ks *keyspace) Has(ctx context.Context, k []byte) (bool, error) {
 	return exists, nil
 }
 
-func (ks *keyspace) Set(ctx context.Context, k, v []byte, r kv.Revision) error {
-	ok, err := ks.set(ctx, v, k, r)
+func (ks *keyspace) Set(ctx context.Context, k, v []byte, r kv.Revision) (kv.Revision, error) {
+	next, ok, err := ks.set(ctx, v, k, r)
 	if ok || err != nil {
-		return err
+		return next, err
 	}
 
-	return kv.ConflictError[[]byte]{
+	return "", kv.ConflictError[[]byte]{
 		Keyspace: ks.name,
 		Key:      k,
 		Revision: r,
 	}
 }
 
-// set inserts, updates, or deletes a key/value pair based on the provided
-// value and revision.
+// set inserts, updates, or deletes a key/value pair based on the provided value
+// and revision.
 //
-// It returns true on success, or false on conflict.
-func (ks *keyspace) set(ctx context.Context, v []byte, k []byte, r kv.Revision) (bool, error) {
+// It returns the new revision and true on success, or empty and false on
+// conflict.
+func (ks *keyspace) set(ctx context.Context, v []byte, k []byte, r kv.Revision) (kv.Revision, bool, error) {
 	isDelete := len(v) == 0
-	isNew := r == 0
+	isNew := r == ""
 
 	if isDelete && isNew {
 		exists, err := ks.Has(ctx, k)
-		return !exists, err
+		return "", !exists, err
+	}
+
+	gen, ok := kvrevision.TryUnmarshalGeneration(r)
+	if !ok {
+		return "", false, nil
 	}
 
 	if isDelete {
-		return ks.execOne(
+		ok, err := ks.execOne(
 			ctx,
 			`DELETE FROM persistencekit.keyspace_pair
 			WHERE keyspace_id = $1
 			AND key = $2
-			AND encoded_revision = $3`,
+			AND encoded_generation = $3`,
 			ks.id,
 			k,
-			bigint.ConvertUnsigned(&r),
+			bigint.ConvertUnsigned(&gen),
 		)
+		return "", ok, err
 	}
 
 	if isNew {
-		return ks.execOne(
+		ok, err := ks.execOne(
 			ctx,
 			`INSERT INTO persistencekit.keyspace_pair AS o (
 				keyspace_id,
@@ -115,21 +124,24 @@ func (ks *keyspace) set(ctx context.Context, v []byte, k []byte, r kv.Revision) 
 			k,
 			v,
 		)
+		return kvrevision.MarshalGeneration(1), ok, err
 	}
 
-	return ks.execOne(
+	ok, err := ks.execOne(
 		ctx,
 		`UPDATE persistencekit.keyspace_pair SET
 			value = $3,
-			encoded_revision = encoded_revision + 1
+			encoded_generation = encoded_generation + 1
 		WHERE keyspace_id = $1
-		AND key = $2
-		AND encoded_revision = $4`,
+			AND key = $2
+			AND encoded_generation = $4`,
 		ks.id,
 		k,
 		v,
-		bigint.ConvertUnsigned(&r),
+		bigint.ConvertUnsigned(&gen),
 	)
+
+	return kvrevision.MarshalGeneration(gen + 1), ok, err
 }
 
 func (ks *keyspace) SetUnconditional(ctx context.Context, k, v []byte) error {
@@ -155,7 +167,7 @@ func (ks *keyspace) SetUnconditional(ctx context.Context, k, v []byte) error {
 			$1, $2, $3
 		) ON CONFLICT (keyspace_id, key) DO UPDATE SET
 			value = EXCLUDED.value,
-			encoded_revision = p.encoded_revision + 1`,
+			encoded_generation = p.encoded_generation + 1`,
 		ks.id,
 		k,
 		v,
@@ -166,7 +178,7 @@ func (ks *keyspace) SetUnconditional(ctx context.Context, k, v []byte) error {
 func (ks *keyspace) Range(ctx context.Context, fn kv.BinaryRangeFunc) error {
 	rows, err := ks.db.QueryContext(
 		ctx,
-		`SELECT key, value, encoded_revision
+		`SELECT key, value, encoded_generation
 		FROM persistencekit.keyspace_pair
 		WHERE keyspace_id = $1`,
 		ks.id,
@@ -179,17 +191,17 @@ func (ks *keyspace) Range(ctx context.Context, fn kv.BinaryRangeFunc) error {
 	for rows.Next() {
 		var (
 			k, v []byte
-			r    kv.Revision
+			gen  uint64
 		)
 		if err := rows.Scan(
 			&k,
 			&v,
-			bigint.ConvertUnsigned(&r),
+			bigint.ConvertUnsigned(&gen),
 		); err != nil {
 			return fmt.Errorf("cannot scan keyspace pair: %w", err)
 		}
 
-		ok, err := fn(ctx, k, v, r)
+		ok, err := fn(ctx, k, v, kvrevision.MarshalGeneration(gen))
 		if !ok || err != nil {
 			return err
 		}
